requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
betterproto==2.0.0b6
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
boto3==1.34.14
botocore==1.34.14
build==1.2.2.post1
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==1.0.7
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.9
filelock==3.16.1
Flask==3.0.0
Flask-SocketIO==5.3.6
flatbuffers==25.2.10
fonttools==4.54.1
fsspec==2023.12.2
gdown==5.2.0
google-auth==2.39.0
google-genai==1.12.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpclib==0.4.7
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
id==1.5.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==5.13.0
in-toto-attestation==0.9.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.9.0
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kaggle==1.6.17
kagglehub==0.3.7
kiwisolver==1.4.7
kubernetes==32.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.1.0
model-signing==0.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
ollama==0.4.8
onnxruntime==1.19.2
openai==1.76.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
posthog==4.0.0
protobuf==5.29.3
psutil==5.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.10.1
python-slugify==8.0.4
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rfc3161-client==0.1.2
rfc8785==0.1.4
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.10.3
scikit-learn==1.3.2
scipy==1.11.4
securesystemslib==1.2.0
shellingham==1.5.4
sigstore==3.6.1
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.45.3
sympy==1.13.1
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.0.2
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
tuf==5.1.0
typer==0.15.2
typing_extensions==4.13.2
tzdata==2024.2
urllib3==1.26.18
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.0.3
zipp==3.20.2
